+++ src/Admin/Infrastructure/Http/ClearHttpCacheController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controller\Admin;
+namespace App\Admin\Infrastructure\Http;
 
 use App\Http\Cache\HttpCache;
 use EasyCorp\Bundle\EasyAdminBundle\Router\AdminUrlGenerator;
@@ -8,7 +8,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 
-final class ClearHttpCache extends AbstractController
+final class ClearHttpCacheController extends AbstractController
 {
     public function __construct(
         private HttpCache $httpCache,
